fix again

+++ db/migrate/20150719205501_time_col_is_dumb.rb
@@ -1,6 +1,6 @@
 class TimeColIsDumb < ActiveRecord::Migration
   def change
     remove_column :tournaments, :time_cap
-    add_column :tournaments, :time_cap, :float
+    add_column :tournaments, :time_cap, :decimal, precision: 3
   end
 end
